Fix support for end to end flows using version 4.2 through 4.5. The issue was that the revocationConfig of the Corda firewall was not set correctly. This would impact users where crl endpoints were not reachable.
Also fix for uninitialized databases for new nodes with (database.runMigration=true).
Unfortunately this parameter should not be enabled for production, but new users may not know how to set up the database properly, so this is still a more reasonable default for people getting familiar with the Corda deployment topic in general.