Add support for database driver handling. Now we automatically download the 3 main supported database drivers (MS SQL, Oracle, Postgres).
The initial registration now also uses a subset of the node.conf, the bare necessities for registering a node onto a network.
copy-cordapps.sh was renamed to copy-files.sh, and now also copies the database drivers into persistent storage.
Note that it is not yet automatically called from helm_compile.sh, due to being slightly slow, need to figure out where to call it in an automated fashion.